Remove false positive GHSA-8x6c-cv3v-vp6g
It is not generally valuable to track vulnerabilities in the dependency
of a project since this (a) duplicates reporting of issues that should
already be tracked in the dependency, and (b) is challenging to do
accurately since some package managers like Yarn allow a project to
override the dependency version. In this particular case,
cacheable-request hasn't pinned a vulnerable version of
http-cache-semantics since v2. Since v3, cacheable-request has pinned
http-cache-semantics to ^4.0.0 or ^4.1.0, both of which include patched
version 4.1.1.