[GHSA-c5q2-7r4c-mv6g]: add missing package to the list of affected packages
As reported in https://pkg.go.dev/vuln/GO-2024-2631, the go module `gopkg.in/square/go-jose.v2` is also affected by this reported vulnerability. Despite of been a deprecated repository, there is still a lot of software using it. So it'd be great to add it to this list of affected packages.